Added ability to specify the intersection point in alarms
(cherry picked from commit b0f36559b5c1bf8a950658d2d6a64331f3bd6c14)

Historically, we've used the bottom middle point of a detection when
deciding if a detection is within a zone. This feature allows users to
specify the point to be the top, bottom, left, right, or center point
when creating alarms.

- Added the "intersection_point" field to the REST API
- Added the "intersection_point" field to the database
- Added database migration
- Added support for this intersection point in the traffic algorithms