Add a better failure mode when the user attempts to do too much analysis
Before, when the user adds more streams than their license allows, the
client would report a somewhat unhelpful message and keep the stream
there. Since the client waits for analysis results before showing a
stream, the stream will never appear to connect. It was pretty messy.
Now, a better message is displayed and the stream is automatically
deleted.