Start the StreamReader inside the UI thread instead of a QThread
This fixes BF-997, where we weren't getting messages from GStreamer
pipelines and were therefor unable to restart them in case of an error.
This was because we were creating the pipeline in a QThread, which as
its own GLib.MainContext, and the context's events aren't processed in
the background since that's not how Qt does things. Luckily, creating a
StreamReader is a non-blocking operation anyway, so now we're just
running the requisite API requests in a QThread and creating the
StreamReader in the main thread.

Also, the UI no longer starts up its own GLib.MainLoop and reuses Qt's
event handling mechanism for the default GLib.MainContext instead. This
doesn't seem to have a big effect on anything in practice, but is
technically more correct.