Remove Multichef in favor of ThreadPoolExecutor (and bump vcap) (#6)
Remove Multichef in favor of ThreadPoolExecutor (and bump vcap)

This removes the idea of "Recipes", "Ovens", or "RecipeRunners". The "Oven" persists but is renamed to "BatchExecutor" in the vcap repository.

Importantly, with these changes, there shouldn't be any kind of exception that could break the inference system.

If an exception happens in the BatchExecutor, it is now re-raised in process_frame, which is caught and logged. Before, it would stop the entire BatchExecutor thread permanently.

Fixes BF-1432