Add Snap support to the client (#1)
This allows users to install the BrainFrame client as a Snap. The self-contained nature of the Snap allows it to work on a majority of Linux distributions, but unfortunately segfaults on Manjaro and likely Arch Linux. The Snap also includes a Qt-native dark theme and provides a desktop icon when installed.

There are currently some limitations to the Snap, like its use of an old GStreamer version (1.14) and lack of VA-API support. Additions to the makedeb project should make it possible to address these shortcomings in the future. The total installation size is also larger than the traditional client distribution. It might be possible to narrow this gap with some effort, but the traditional distribution will probably always be smaller.

Fixes BF-1439